Move schema code into its own module
