+++ src/test/complex_dependency_test.ts
@@ -2,6 +2,7 @@
 import * as test  from "https://deno.land/std@0.208.0/assert/mod.ts";
 import * as path  from "https://deno.land/std@0.208.0/path/mod.ts";
 import * as infer from '../infer.ts';
+import { FunctionNdcKind } from "../schema.ts";
 
 // This program omits its return type and it is inferred via the 'fetch' dependency.
 Deno.test("Inference on Dependency", () => {
@@ -10,7 +11,7 @@ Deno.test("Inference on Dependency", () => {
   const program_schema = infer.inferProgramSchema(program_path, vendor_path, false);
   test.assertEquals(program_schema.functions, {
     infinite_loop: {
-      ndc_kind: infer.FunctionNdcKind.Procedure,
+      ndc_kind: FunctionNdcKind.Procedure,
       description: null,
       arguments: [],
       result_type: { type: "named", kind: "object", name: "Response" }
